feat: implement create_multibanco_payment command and enhance payment handling; update argument types and add new payment methods

+++ helpdeskMain/eupago/domain/money.py
@@ -3,7 +3,7 @@ from dataclasses import dataclass
 # Add logger
 @dataclass
 class Money:
-    amount: int # CHECK OUT FLOAT/INT
+    amount: float
     currency: str
 
     def __post_init__(self):
